American College of Radiology Lung Cancer Screening Registry (LCSR) dataset

src_lung_cancer_screening_data() returns a dataframe of actively reporting sites in the lung cancer screening registry (LCSR).

src_acr_lung_cancer_screening_data(state = NULL, ...)

Source

Data Source

Arguments

state

Character string or vector of character strings of state postal abbreviations

...

passed to readr::read_csv() and useful for limiting the number of rows read for testing or glimpsing data.

Value

A data frame (data.frame) containing data pulled from LCSR.

Details

The ACR Lung Cancer Screening Registry (LCSR) is designed to systematically audit the quality of interpretation of screening lung CT exams. The registry is based on the ACR Lung Imaging Reporting and Data System (Lung-RADS), which is the product of the ACR Lung Cancer Screening Committee subgroup on Lung-RADS. This Lung-RADS system is a quality assurance tool designed to standardize lung cancer screening CT reporting and management recommendations, reduce confusion in lung cancer screening CT interpretations and facilitate outcome monitoring. The ACR LCSR will capture Lung-RADS recommendations and monitor and compare appropriate use of Lung-RADS#'

See also

Author

Todd Burus tburus@uky.edu

Examples

# example code

lcs = src_acr_lung_cancer_screening_data(nrows=1000)
#> INFO  [15:23:38.133] Starting acr_lung_cancer_screening_data
#> Error in read_csv("https://report.acr.org/t/PUBLIC/views/NRDRLCSLocator/ACRLCSDownload.csv",     ...): unused argument (nrows = 1000)

colnames(lcs)
#> Error in eval(expr, envir, enclos): object 'lcs' not found

dplyr::glimpse(lcs)
#> Error in eval(expr, envir, enclos): object 'lcs' not found