CDC PLACES

src_cdc_places_data() returns a dataframe of estimates from CDC PLACES.

Usage

src_cdc_places_data(level = "county", year = 2023, ...)

Source

Data Source

Arguments

level

Character string for desired geographic level. Available at 'county', "tract", or 'zcta'. Defaults to 'county'.

year

The year of the CDC PLACES estimates. Estimates are available from 2020 through 2023. Defaults to 2023.

Value

A data frame (data.frame) containing data pulled from the CDC PLACES API.

Details

PLACES is a collaboration between CDC, the Robert Wood Johnson Foundation, and the CDC Foundation. PLACES provides health data for small areas across the country. This allows local health departments and jurisdictions, regardless of population size and rurality, to better understand the burden and geographic distribution of health measures in their areas and assist them in planning public health interventions. PLACES provides model-based, population-level analysis and community estimates of health measures to all counties, places (incorporated and census designated places), census tracts, and ZIP Code Tabulation Areas (ZCTAs) across the United States.

See also

Author

Todd Burus tburus@uky.edu

Examples

# example code

places = src_cdc_places_data(level = 'tract', nrows = 500)
#> INFO  [13:40:53.750] Starting cdc_places_data
#> INFO  [13:40:54.377] Completing cdc_places_data

colnames(places)
#>  [1] "StateAbbr"              "StateDesc"              "CountyName"            
#>  [4] "CountyFIPS"             "TractFIPS"              "TotalPopulation"       
#>  [7] "ACCESS2_CrudePrev"      "ACCESS2_Crude95CI"      "ARTHRITIS_CrudePrev"   
#> [10] "ARTHRITIS_Crude95CI"    "BINGE_CrudePrev"        "BINGE_Crude95CI"       
#> [13] "BPHIGH_CrudePrev"       "BPHIGH_Crude95CI"       "BPMED_CrudePrev"       
#> [16] "BPMED_Crude95CI"        "CANCER_CrudePrev"       "CANCER_Crude95CI"      
#> [19] "CASTHMA_CrudePrev"      "CASTHMA_Crude95CI"      "CERVICAL_CrudePrev"    
#> [22] "CERVICAL_Crude95CI"     "CHD_CrudePrev"          "CHD_Crude95CI"         
#> [25] "CHECKUP_CrudePrev"      "CHECKUP_Crude95CI"      "CHOLSCREEN_CrudePrev"  
#> [28] "CHOLSCREEN_Crude95CI"   "COLON_SCREEN_CrudePrev" "COLON_SCREEN_Crude95CI"
#> [31] "COPD_CrudePrev"         "COPD_Crude95CI"         "COREM_CrudePrev"       
#> [34] "COREM_Crude95CI"        "COREW_CrudePrev"        "COREW_Crude95CI"       
#> [37] "CSMOKING_CrudePrev"     "CSMOKING_Crude95CI"     "DENTAL_CrudePrev"      
#> [40] "DENTAL_Crude95CI"       "DEPRESSION_CrudePrev"   "DEPRESSION_Crude95CI"  
#> [43] "DIABETES_CrudePrev"     "DIABETES_Crude95CI"     "GHLTH_CrudePrev"       
#> [46] "GHLTH_Crude95CI"        "HIGHCHOL_CrudePrev"     "HIGHCHOL_Crude95CI"    
#> [49] "KIDNEY_CrudePrev"       "KIDNEY_Crude95CI"       "LPA_CrudePrev"         
#> [52] "LPA_Crude95CI"          "MAMMOUSE_CrudePrev"     "MAMMOUSE_Crude95CI"    
#> [55] "MHLTH_CrudePrev"        "MHLTH_Crude95CI"        "OBESITY_CrudePrev"     
#> [58] "OBESITY_Crude95CI"      "PHLTH_CrudePrev"        "PHLTH_Crude95CI"       
#> [61] "SLEEP_CrudePrev"        "SLEEP_Crude95CI"        "STROKE_CrudePrev"      
#> [64] "STROKE_Crude95CI"       "TEETHLOST_CrudePrev"    "TEETHLOST_Crude95CI"   
#> [67] "HEARING_CrudePrev"      "HEARING_Crude95CI"      "VISION_CrudePrev"      
#> [70] "VISION_Crude95CI"       "COGNITION_CrudePrev"    "COGNITION_Crude95CI"   
#> [73] "MOBILITY_CrudePrev"     "MOBILITY_Crude95CI"     "SELFCARE_CrudePrev"    
#> [76] "SELFCARE_Crude95CI"     "INDEPLIVE_CrudePrev"    "INDEPLIVE_Crude95CI"   
#> [79] "DISABILITY_CrudePrev"   "DISABILITY_Crude95CI"   "Geolocation"           

dplyr::glimpse(places)
#> Rows: 500
#> Columns: 81
#> $ StateAbbr              <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL",…
#> $ StateDesc              <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Al…
#> $ CountyName             <chr> "Autauga", "Autauga", "Autauga", "Autauga", "Au…
#> $ CountyFIPS             <int> 1001, 1001, 1001, 1001, 1001, 1001, 1001, 1001,…
#> $ TractFIPS              <int> 1001020100, 1001020200, 1001020300, 1001020400,…
#> $ TotalPopulation        <int> 1912, 2170, 3373, 4386, 10766, 3668, 2891, 3081…
#> $ ACCESS2_CrudePrev      <dbl> 10.2, 13.7, 11.4, 7.9, 8.4, 13.2, 14.6, 8.0, 9.…
#> $ ACCESS2_Crude95CI      <chr> "( 7.6, 13.1)", "(11.0, 16.8)", "( 8.9, 14.6)",…
#> $ ARTHRITIS_CrudePrev    <dbl> 30.1, 28.8, 30.1, 32.0, 26.5, 31.3, 32.2, 28.8,…
#> $ ARTHRITIS_Crude95CI    <chr> "(25.9, 34.6)", "(24.6, 32.9)", "(26.0, 34.4)",…
#> $ BINGE_CrudePrev        <dbl> 15.4, 13.9, 15.1, 15.0, 16.3, 14.3, 14.2, 16.4,…
#> $ BINGE_Crude95CI        <chr> "(12.5, 18.6)", "(11.3, 16.9)", "(12.2, 18.3)",…
#> $ BPHIGH_CrudePrev       <dbl> 37.7, 42.2, 38.5, 38.3, 33.7, 40.8, 40.7, 36.2,…
#> $ BPHIGH_Crude95CI       <chr> "(33.3, 42.5)", "(37.6, 46.8)", "(34.1, 42.9)",…
#> $ BPMED_CrudePrev        <dbl> 79.0, 78.8, 79.1, 81.7, 76.6, 79.2, 79.0, 78.7,…
#> $ BPMED_Crude95CI        <chr> "(75.7, 82.2)", "(75.8, 81.7)", "(75.9, 82.0)",…
#> $ CANCER_CrudePrev       <dbl> 6.6, 5.3, 6.6, 8.4, 6.2, 6.2, 6.4, 6.5, 6.1, 6.…
#> $ CANCER_Crude95CI       <chr> "( 5.9,  7.3)", "( 4.8,  5.8)", "( 5.9,  7.3)",…
#> $ CASTHMA_CrudePrev      <dbl> 10.3, 11.3, 10.6, 9.3, 9.8, 11.0, 11.5, 9.6, 10…
#> $ CASTHMA_Crude95CI      <chr> "( 8.9, 11.7)", "( 9.9, 12.9)", "( 9.3, 12.1)",…
#> $ CERVICAL_CrudePrev     <dbl> 83.5, 84.6, 82.6, 85.9, 86.0, 82.3, 80.2, 85.6,…
#> $ CERVICAL_Crude95CI     <chr> "(81.0, 86.0)", "(82.8, 86.4)", "(80.2, 85.0)",…
#> $ CHD_CrudePrev          <dbl> 6.0, 5.4, 6.1, 6.5, 4.7, 6.9, 7.1, 5.3, 5.5, 5.…
#> $ CHD_Crude95CI          <chr> "( 5.2,  6.8)", "( 4.8,  6.2)", "( 5.3,  7.0)",…
#> $ CHECKUP_CrudePrev      <dbl> 75.8, 77.7, 76.0, 77.5, 75.5, 75.9, 75.3, 75.8,…
#> $ CHECKUP_Crude95CI      <chr> "(70.7, 80.2)", "(73.1, 82.0)", "(71.0, 80.5)",…
#> $ CHOLSCREEN_CrudePrev   <dbl> 85.2, 84.2, 84.5, 88.4, 86.5, 83.1, 81.6, 87.3,…
#> $ CHOLSCREEN_Crude95CI   <chr> "(82.1, 87.8)", "(81.2, 86.9)", "(81.4, 87.4)",…
#> $ COLON_SCREEN_CrudePrev <dbl> 70.8, 70.9, 70.8, 75.5, 72.7, 68.2, 66.5, 73.0,…
#> $ COLON_SCREEN_Crude95CI <chr> "(67.5, 73.9)", "(68.0, 73.6)", "(67.6, 73.9)",…
#> $ COPD_CrudePrev         <dbl> 7.6, 7.4, 7.8, 6.8, 5.6, 9.1, 10.3, 6.3, 7.1, 6…
#> $ COPD_Crude95CI         <chr> "( 6.2,  9.2)", "( 6.1,  8.6)", "( 6.5,  9.3)",…
#> $ COREM_CrudePrev        <dbl> 44.0, 37.5, 42.4, 49.8, 47.3, 40.5, 39.1, 45.9,…
#> $ COREM_Crude95CI        <chr> "(35.0, 53.7)", "(30.8, 44.5)", "(34.6, 51.4)",…
#> $ COREW_CrudePrev        <dbl> 42.8, 37.4, 40.5, 46.3, 44.1, 37.5, 36.8, 44.5,…
#> $ COREW_Crude95CI        <chr> "(34.8, 50.6)", "(31.7, 43.6)", "(33.2, 48.1)",…
#> $ CSMOKING_CrudePrev     <dbl> 18.0, 20.6, 18.7, 13.5, 14.3, 21.7, 24.0, 15.3,…
#> $ CSMOKING_Crude95CI     <chr> "(14.2, 21.8)", "(16.7, 24.8)", "(14.9, 22.6)",…
#> $ DENTAL_CrudePrev       <dbl> 63.4, 55.6, 61.1, 70.3, 68.4, 56.1, 53.5, 68.4,…
#> $ DENTAL_Crude95CI       <chr> "(59.5, 67.4)", "(52.4, 58.8)", "(57.3, 64.7)",…
#> $ DEPRESSION_CrudePrev   <dbl> 23.8, 21.3, 23.9, 21.8, 22.6, 24.8, 26.4, 22.5,…
#> $ DEPRESSION_Crude95CI   <chr> "(19.8, 28.4)", "(17.6, 25.3)", "(19.8, 28.1)",…
#> $ DIABETES_CrudePrev     <dbl> 11.0, 13.7, 11.5, 10.3, 8.9, 13.2, 13.3, 9.9, 1…
#> $ DIABETES_Crude95CI     <chr> "( 9.3, 12.9)", "(11.7, 15.7)", "( 9.7, 13.2)",…
#> $ GHLTH_CrudePrev        <dbl> 17.9, 22.4, 19.1, 14.7, 14.0, 22.4, 24.1, 14.9,…
#> $ GHLTH_Crude95CI        <chr> "(14.6, 21.1)", "(19.1, 25.8)", "(16.1, 22.2)",…
#> $ HIGHCHOL_CrudePrev     <dbl> 36.0, 33.0, 35.3, 37.9, 33.1, 36.2, 36.7, 35.3,…
#> $ HIGHCHOL_Crude95CI     <chr> "(31.3, 40.8)", "(28.7, 37.7)", "(30.9, 39.9)",…
#> $ KIDNEY_CrudePrev       <dbl> 2.9, 3.2, 3.0, 3.1, 2.5, 3.2, 3.3, 2.6, 2.7, 2.…
#> $ KIDNEY_Crude95CI       <chr> "( 2.6,  3.2)", "( 2.9,  3.5)", "( 2.7,  3.4)",…
#> $ LPA_CrudePrev          <dbl> 29.6, 35.3, 31.5, 25.7, 24.7, 34.7, 37.3, 25.4,…
#> $ LPA_Crude95CI          <chr> "(24.4, 35.5)", "(30.1, 41.3)", "(26.1, 37.0)",…
#> $ MAMMOUSE_CrudePrev     <dbl> 76.8, 81.0, 76.9, 75.5, 77.8, 77.5, 76.0, 78.5,…
#> $ MAMMOUSE_Crude95CI     <chr> "(72.6, 80.6)", "(78.3, 83.7)", "(73.3, 80.6)",…
#> $ MHLTH_CrudePrev        <dbl> 18.3, 19.4, 18.9, 15.2, 16.6, 20.3, 21.8, 16.4,…
#> $ MHLTH_Crude95CI        <chr> "(15.6, 20.9)", "(16.9, 22.1)", "(16.3, 21.5)",…
#> $ OBESITY_CrudePrev      <dbl> 38.7, 45.0, 39.4, 34.8, 36.3, 42.6, 42.6, 37.3,…
#> $ OBESITY_Crude95CI      <chr> "(31.4, 46.3)", "(37.6, 52.7)", "(32.2, 46.9)",…
#> $ PHLTH_CrudePrev        <dbl> 12.5, 13.2, 12.8, 10.9, 10.1, 14.6, 15.7, 10.9,…
#> $ PHLTH_Crude95CI        <chr> "(10.6, 14.7)", "(11.4, 15.3)", "(10.9, 14.8)",…
#> $ SLEEP_CrudePrev        <dbl> 36.9, 43.4, 38.1, 33.4, 35.7, 39.5, 39.8, 35.8,…
#> $ SLEEP_Crude95CI        <chr> "(35.1, 38.6)", "(42.0, 44.8)", "(36.6, 39.6)",…
#> $ STROKE_CrudePrev       <dbl> 3.0, 3.6, 3.3, 3.1, 2.4, 3.6, 3.7, 2.7, 2.8, 2.…
#> $ STROKE_Crude95CI       <chr> "( 2.6,  3.5)", "( 3.2,  4.0)", "( 2.9,  3.7)",…
#> $ TEETHLOST_CrudePrev    <dbl> 14.6, 19.4, 16.2, 10.9, 11.1, 19.3, 21.5, 11.2,…
#> $ TEETHLOST_Crude95CI    <chr> "( 9.7, 20.1)", "(14.4, 25.0)", "(11.3, 21.5)",…
#> $ HEARING_CrudePrev      <dbl> 6.3, 5.0, 6.4, 6.9, 5.2, 7.0, 7.4, 5.6, 5.8, 5.…
#> $ HEARING_Crude95CI      <chr> "( 5.5,  7.2)", "( 4.4,  5.6)", "( 5.6,  7.2)",…
#> $ VISION_CrudePrev       <dbl> 4.7, 6.5, 5.2, 3.6, 3.6, 6.6, 7.1, 3.7, 4.4, 3.…
#> $ VISION_Crude95CI       <chr> "( 4.0,  5.5)", "( 5.6,  7.5)", "( 4.5,  6.0)",…
#> $ COGNITION_CrudePrev    <dbl> 14.6, 16.7, 15.6, 11.4, 12.2, 17.7, 19.6, 12.2,…
#> $ COGNITION_Crude95CI    <chr> "(12.1, 17.2)", "(14.4, 19.3)", "(13.2, 18.3)",…
#> $ MOBILITY_CrudePrev     <dbl> 15.5, 17.8, 16.4, 14.2, 12.0, 18.8, 20.2, 13.0,…
#> $ MOBILITY_Crude95CI     <chr> "(12.9, 18.2)", "(15.4, 20.6)", "(13.9, 19.0)",…
#> $ SELFCARE_CrudePrev     <dbl> 3.9, 5.0, 4.1, 3.0, 2.9, 5.3, 5.6, 3.1, 3.7, 3.…
#> $ SELFCARE_Crude95CI     <chr> "( 3.3,  4.5)", "( 4.4,  5.8)", "( 3.5,  4.7)",…
#> $ INDEPLIVE_CrudePrev    <dbl> 8.3, 10.3, 9.1, 6.6, 6.7, 10.8, 11.9, 6.7, 7.9,…
#> $ INDEPLIVE_Crude95CI    <chr> "( 7.0,  9.8)", "( 8.8, 11.8)", "( 7.7, 10.6)",…
#> $ DISABILITY_CrudePrev   <dbl> 30.9, 33.8, 32.5, 28.3, 25.9, 35.8, 38.3, 26.8,…
#> $ DISABILITY_Crude95CI   <chr> "(26.0, 35.8)", "(29.0, 38.5)", "(27.8, 37.4)",…
#> $ Geolocation            <chr> "POINT (-86.49001486 32.47718251)", "POINT (-86…