US CDC Social Vulnerability Index

The CDC's Social Vulnerability Index (SVI), created and maintained by the Geospatial Research, Analysis, and Services Program (GRASP), uses US Census data to determine the social vulnerability of every county and tract. This index ranks each county and tract based upon 15 social factors including poverty, lack of vehicle access, and crowded housing, and groups them into four related themes: 1) Socioeconomic, 2) Housing Composition and Disability, 3) Minority Status and Language, and 4) Housing and Transportation.

Usage

src_cdc_social_vulnerability_index()

Source

https://svi.cdc.gov/

Value

a data.frame

Details

Theme rankings: For each of the four themes, we summed the percentiles for the variables comprising each theme. We ordered the summed percentiles for each theme to determine theme-specific percentile rankings. The four summary theme ranking variables, detailed in the Data Dictionary below, are:

Socioeconomic - RPL_THEME1

Household Composition & Disability - RPL_THEME2

Minority Status & Language - RPL_THEME3

Housing Type & Transportation - RPL_THEME4

Overall tract rankings: We summed the sums for each theme, ordered the tracts, and then calculated overall percentile rankings. Please note; taking the sum of the sums for each theme is the same as summing individual variable rankings. The overall tract summary ranking variable is RPL_THEM

For detailed documentation, see https://svi.cdc.gov/Documents/Data/2018_SVI_Data/SVI2018Documentation.pdf

Author

Sean Davis seandavi@gmail.com

Examples

res = src_cdc_social_vulnerability_index()
head(res)
#> # A tibble: 6 × 126
#>   grasp_id state_fips state   st_abbr county  fips  location  area_sqmi e_totpop
#>      <dbl> <chr>      <chr>   <chr>   <chr>   <chr> <chr>         <dbl>    <dbl>
#> 1        1 00001      Alabama AL      Autauga 01001 Autauga …      594.    55200
#> 2        2 00001      Alabama AL      Baldwin 01003 Baldwin …     1590.   208107
#> 3        3 00001      Alabama AL      Barbour 01005 Barbour …      885.    25782
#> 4        4 00001      Alabama AL      Bibb    01007 Bibb Cou…      622.    22527
#> 5        5 00001      Alabama AL      Blount  01009 Blount C…      645.    57645
#> 6        6 00001      Alabama AL      Bullock 01011 Bullock …      623.    10352
#> # ℹ 117 more variables: m_totpop <dbl>, e_hu <dbl>, m_hu <dbl>, e_hh <dbl>,
#> #   m_hh <dbl>, e_pov <dbl>, m_pov <dbl>, e_unemp <dbl>, m_unemp <dbl>,
#> #   e_pci <dbl>, m_pci <dbl>, e_nohsdp <dbl>, m_nohsdp <dbl>, e_age65 <dbl>,
#> #   m_age65 <dbl>, e_age17 <dbl>, m_age17 <dbl>, e_disabl <dbl>,
#> #   m_disabl <dbl>, e_sngpnt <dbl>, m_sngpnt <dbl>, e_minrty <dbl>,
#> #   m_minrty <dbl>, e_limeng <dbl>, m_limeng <dbl>, e_munit <dbl>,
#> #   m_munit <dbl>, e_mobile <dbl>, m_mobile <dbl>, e_crowd <dbl>, …

# limit to index columns only 
res |> dplyr::select(
    state_fips:e_totpop,dplyr::starts_with('rpl_'))
#> # A tibble: 3,142 × 13
#>    state_fips state  st_abbr county fips  location area_sqmi e_totpop rpl_theme1
#>    <chr>      <chr>  <chr>   <chr>  <chr> <chr>        <dbl>    <dbl>      <dbl>
#>  1 00001      Alaba… AL      Autau… 01001 Autauga…      594.    55200      0.363
#>  2 00001      Alaba… AL      Baldw… 01003 Baldwin…     1590.   208107      0.223
#>  3 00001      Alaba… AL      Barbo… 01005 Barbour…      885.    25782      0.978
#>  4 00001      Alaba… AL      Bibb   01007 Bibb Co…      622.    22527      0.769
#>  5 00001      Alaba… AL      Blount 01009 Blount …      645.    57645      0.614
#>  6 00001      Alaba… AL      Bullo… 01011 Bullock…      623.    10352      0.977
#>  7 00001      Alaba… AL      Butler 01013 Butler …      777.    20025      0.846
#>  8 00001      Alaba… AL      Calho… 01015 Calhoun…      606.   115098      0.787
#>  9 00001      Alaba… AL      Chamb… 01017 Chamber…      597.    33826      0.690
#> 10 00001      Alaba… AL      Chero… 01019 Cheroke…      554.    25853      0.688
#> # ℹ 3,132 more rows
#> # ℹ 4 more variables: rpl_theme2 <dbl>, rpl_theme3 <dbl>, rpl_theme4 <dbl>,
#> #   rpl_themes <dbl>