National Plan & Provider Enumeration System (NPPES)

src_nppes_onco_data() returns a dataframe of oncologists from the NPPES API.

Usage

src_nppes_onco_data(state = c(state.abb, "DC"))

Source

Data Source

Arguments

state

Character string or vector of character strings of state postal abbreviations

Value

A data frame (data.frame) containing data pulled from the NPPES API.

Details

Centers for Medicare & Medicaid Services CMS has developed the NPPES to assign unique identifiers to health care providers. The National Provider Indentifier (NPI) has been the standard identifier for all HIPAA-covered entities (health care providers) since May 23, 2007. Small health plans were required to obtain and use an NPI by May 23, 2008.

See also

Author

Todd Burus tburus@uky.edu

Examples

# example code

onc = src_nppes_onco_data(state  = c('KY', 'TN'))
#> INFO  [13:41:18.105] Starting nppes_onco_data
#> Collecting data for KY 
#> Collecting data for TN 
#> INFO  [13:41:22.706] Completing nppes_onco_data

colnames(onc)
#> [1] "Type"         "Name"         "Address"      "State"        "Phone_number"
#> [6] "Notes"        "latitude"     "longitude"   

dplyr::glimpse(onc)
#> Rows: 1,355
#> Columns: 8
#> $ Type         <chr> "Oncologist", "Oncologist", "Oncologist", "Oncologist", "…
#> $ Name         <chr> "Mehdi Ahmadi", "Mohamed Alsharedi", "Lowell Anthony", "R…
#> $ Address      <chr> "204 South 9th St, Murray, KY 42071", "2195 Harrodsburg R…
#> $ State        <chr> "KY", "KY", "KY", "KY", "FL", "KY", "KY", "AZ", "KY", "KY…
#> $ Phone_number <chr> "270-759-4199", "859-258-4673", "859-323-6522", "859-323-…
#> $ Notes        <chr> "Medical Oncology", "Medical Oncology", "Medical Oncology…
#> $ latitude     <chr> "", "", "", "", "", "", "", "", "", "", "", "", "", "", "…
#> $ longitude    <chr> "", "", "", "", "", "", "", "", "", "", "", "", "", "", "…