Behavioral Risk Factor Surveillance System (BRFSS) dataset

The Behavioral Risk Factor Surveillance System (BRFSS) is the nation’s premier system of health-related telephone surveys that collect state data about U.S. residents regarding their health-related risk behaviors, chronic health conditions, and use of preventive services. Established in 1984 with 15 states, BRFSS now collects data in all 50 states as well as the District of Columbia and three U.S. territories. BRFSS completes more than 400,000 adult interviews each year, making it the largest continuously conducted health survey system in the world.

src_brfss(...)

Source

Data Source

Arguments

...

passed to readr::read_csv() and useful for limiting the number of rows read for testing or glimpsing data.

Details

Fact sheet is available here.

See also

Author

Sean Davis seandavi@gmail.com

Examples

brfss = src_brfss(n_max=1000)
#> Rows: 1000 Columns: 27
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (21): Locationabbr, Locationdesc, Class, Topic, Question, Response, Brea...
#> dbl  (6): Year, Sample_Size, Data_value, Confidence_limit_Low, Confidence_li...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

colnames(brfss)
#>  [1] "Year"                       "Locationabbr"              
#>  [3] "Locationdesc"               "Class"                     
#>  [5] "Topic"                      "Question"                  
#>  [7] "Response"                   "Break_Out"                 
#>  [9] "Break_Out_Category"         "Sample_Size"               
#> [11] "Data_value"                 "Confidence_limit_Low"      
#> [13] "Confidence_limit_High"      "Display_order"             
#> [15] "Data_value_unit"            "Data_value_type"           
#> [17] "Data_Value_Footnote_Symbol" "Data_Value_Footnote"       
#> [19] "DataSource"                 "ClassId"                   
#> [21] "TopicId"                    "LocationID"                
#> [23] "BreakoutID"                 "BreakOutCategoryID"        
#> [25] "QuestionID"                 "ResponseID"                
#> [27] "GeoLocation"               

dplyr::glimpse(brfss)
#> Rows: 1,000
#> Columns: 27
#> $ Year                       <dbl> 2012, 2012, 2022, 2022, 2022, 2013, 2013, 2…
#> $ Locationabbr               <chr> "GU", "GU", "AL", "AL", "AL", "GU", "GU", "…
#> $ Locationdesc               <chr> "Guam", "Guam", "Alabama", "Alabama", "Alab…
#> $ Class                      <chr> "Demographics", "Demographics", "Alcohol Co…
#> $ Topic                      <chr> "Age", "Age", "Alcohol Consumption", "Alcoh…
#> $ Question                   <chr> "What is your age?", "What is your age?", "…
#> $ Response                   <chr> "55-64 years", "35-44 years", "Yes", "No", …
#> $ Break_Out                  <chr> "Overall", "Female", "18-24", "18-24", "25-…
#> $ Break_Out_Category         <chr> "Overall", "Gender", "Age Group", "Age Grou…
#> $ Sample_Size                <dbl> 296, 272, 95, 84, 249, 69, 48, 183, 149, 29…
#> $ Data_value                 <dbl> 13.1, 21.2, 52.4, 47.6, 59.0, 24.5, 7.8, 20…
#> $ Confidence_limit_Low       <dbl> 11.4, 18.1, 43.7, 38.9, 52.9, 18.4, 5.2, 17…
#> $ Confidence_limit_High      <dbl> 14.9, 24.2, 61.1, 56.3, 65.2, 30.7, 10.5, 2…
#> $ Display_order              <dbl> 1167, 1174, 1175, 1176, 1177, 16, 96, 98, 1…
#> $ Data_value_unit            <chr> "%", "%", "%", "%", "%", "%", "%", "%", "%"…
#> $ Data_value_type            <chr> "Crude Prevalence", "Crude Prevalence", "Cr…
#> $ Data_Value_Footnote_Symbol <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ Data_Value_Footnote        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ DataSource                 <chr> "BRFSS", "BRFSS", "BRFSS", "BRFSS", "BRFSS"…
#> $ ClassId                    <chr> "CLASS05", "CLASS05", "CLASS01", "CLASS01",…
#> $ TopicId                    <chr> "Topic02", "Topic02", "TOPIC03", "TOPIC03",…
#> $ LocationID                 <chr> "66", "66", "01", "01", "01", "66", "66", "…
#> $ BreakoutID                 <chr> "BO1", "SEX2", "AGE01", "AGE01", "AGE02", "…
#> $ BreakOutCategoryID         <chr> "CAT1", "CAT2", "CAT3", "CAT3", "CAT3", "CA…
#> $ QuestionID                 <chr> "AGE", "AGE", "DRNKANY6", "DRNKANY6", "DRNK…
#> $ ResponseID                 <chr> "RESP014", "RESP012", "RESP046", "RESP054",…
#> $ GeoLocation                <chr> "(13.444304, 144.793731)", "(13.444304, 144…